# gin-backend/docs/swagger.yaml
basePath: /
definitions:
  core.ErrorDetail:
    properties:
      code:
        type: string
      details: {}
      message:
        type: string
    type: object
  core.ErrorResponse:
    properties:
      error:
        $ref: '#/definitions/core.ErrorDetail'
    type: object
...
  domain.CreateDomicilioRequest:
    properties:
      alias:
        example: Casa
        type: string
      calle:
        example: Av. Segunda Sur Ote.
        type: string
      colonia_id:
        example: 1
        type: integer
      direccion:
        example: Av. Segunda Sur Ote. 448, Centro, Suchiapa
        type: string
      numero:
        example: "448"
        type: string
      referencia:
        example: Frente al parque
        type: string
      usuario_id:
        example: 100
        type: integer
    type: object
...
  main.CreateUsuarioRequest:
    properties:
      email:
        example: conductor@recolecta.mx
        type: string
      nombre:
        example: Juan Conductor
        type: string
      password:
        example: Conductor123
        type: string
      rol_id:
        example: 4
        type: integer
    type: object
  main.LoginUsuarioRequest:
    properties:
      email:
        example: conductor@recolecta.mx
        type: string
      password:
        example: Conductor123
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: API para gestión de recolección de residuos
  title: API Recolecta
  version: "1.0"
paths:
...
  /api/domicilios:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/core.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/core.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Listar domicilios del usuario autenticado
      tags:
      - Domicilio
    post:
      consumes:
      - application/json
      description: Requiere colonia_id, alias y direccion (o calle). usuario_id es
        opcional si hay JWT.
      parameters:
      - description: Datos del domicilio
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/domain.CreateDomicilioRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/core.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/core.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/core.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Crear domicilio
      tags:
      - Domicilio
  /api/domicilios/{id}:
    delete:
      parameters:
      - description: ID del domicilio
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Eliminar domicilio
      tags:
      - Domicilio
    get:
      parameters:
      - description: ID del domicilio
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/core.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Obtener domicilio por ID
      tags:
      - Domicilio
    put:
      consumes:
      - application/json
      parameters:
      - description: ID del domicilio
        in: path
        name: id
        required: true
        type: integer
      - description: Datos del domicilio
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/domain.CreateDomicilioRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Actualizar domicilio
      tags:
      - Domicilio
...
  /api/usuarios:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Listar usuarios
      tags:
      - Usuario
    post:
      consumes:
      - application/json
      description: 'rol_id: 1=Admin, 2=Coordinador, 3=Operador, 4=Conductor, 5=Ciudadano'
      parameters:
      - description: Datos del usuario
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/main.CreateUsuarioRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      summary: Crear usuario
      tags:
      - Usuario
...
  /api/usuarios/login:
    post:
      consumes:
      - application/json
      parameters:
      - description: Credenciales
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/main.LoginUsuarioRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      summary: Login de usuario/empleado
      tags:
      - Usuario
...
  /domicilios:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Listar domicilios (compat)
      tags:
      - Domicilio
    post:
      consumes:
      - application/json
      parameters:
      - description: Payload JSON
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/domain.CreateDomicilioRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Crear domicilio (compat)
      tags:
      - Domicilio
  /domicilios/{id}:
    delete:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Eliminar domicilio (compat)
      tags:
      - Domicilio
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Obtener domicilio (compat)
      tags:
      - Domicilio
    put:
      consumes:
      - application/json
      parameters:
      - description: Payload JSON
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/domain.CreateDomicilioRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Actualizar domicilio (compat)
      tags:
      - Domicilio
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: 'Ingresa el token JWT con el prefijo Bearer. Ejemplo: Bearer eyJhbGciOiJIUzI1NiIs...'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"